gateway/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
httpx==0.28.1
python-dotenv==1.1.1
